Prep for refactor with databases.
