+++ package.json
@@ -18,7 +18,9 @@
     "dotenv": "^8.0.0",
     "express": "^4.17.1",
     "helmet": "^3.20.0",
+    "knex": "^0.19.1",
     "morgan": "^1.9.1",
+    "pg": "^7.12.1",
     "uuid": "^3.3.2",
     "winston": "^3.2.1"
   },
@@ -26,6 +28,7 @@
     "chai": "^4.2.0",
     "mocha": "^6.2.0",
     "nodemon": "^1.19.1",
+    "postgrator-cli": "^3.1.0",
     "supertest": "^4.0.2"
   }
 }
