+++ postgrator-config.js
@@ -1,0 +1,11 @@
+require('dotenv').config();
+
+module.exports = {
+  "migrationDirectory": "migrations",
+  "driver": "pg",
+  "host": process.env.MIGRATION_DB_HOST,
+  "port": process.env.MIGRATION_DB_PORT,
+  "database": process.env.MIGRATION_DB_NAME,
+  "username": process.env.MIGRATION_DB_USER,
+  "password": process.env.MIGRATION_DB_PASS
+}
