+++ src/app.js
@@ -4,7 +4,7 @@ const morgan = require('morgan')
 const cors = require('cors')
 const helmet = require('helmet')
 const { NODE_ENV } = require('./config')
-const bookmarksRouter = require('./bookmarks/bookmarks-router')
+const bookmarksRouter = require('../bookmarks/bookmarks-router')
 
 const app = express()
